cpt8/8-2_131_palindrome_partitioning_test.go
// https://leetcode-cn.com/problems/palindrome-partitioning/
// 131. 分割回文串
// 输入: "aab"
// 输出:
// [
//   ["aa","b"],
//   ["a","a","b"]
// ]
package cpt8

import (
	"reflect"
	"testing"
)

func Test_partition(t *testing.T) {
	type args struct {
		s string
	}
	tests := []struct {
		name string
		args args
		want [][]string
	}{
		{name: "test1", args: args{s: "aab"}, want: [][]string{{"a", "a", "b"}, {"aa", "b"}}},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := partition(tt.args.s); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("partition() = %v, want %v", got, tt.want)
			}
		})
	}
}
